+++ inputvars/unparsed_value.go
@@ -1,11 +1,20 @@
 package inputvars
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
 	"github.com/hashicorp/hcl/v2"
+	"github.com/spf13/viper"
+	"github.com/turbot/pipe-fittings/backend"
+	"github.com/turbot/pipe-fittings/connection"
+	"github.com/turbot/pipe-fittings/constants"
+	"github.com/turbot/pipe-fittings/error_helpers"
+	"github.com/turbot/pipe-fittings/hclhelpers"
 	"github.com/turbot/pipe-fittings/modconfig"
+	"github.com/turbot/pipe-fittings/pipes"
+	"github.com/turbot/pipe-fittings/steampipeconfig"
 	"github.com/turbot/terraform-components/terraform"
 	"github.com/turbot/terraform-components/tfdiags"
 	"github.com/zclconf/go-cty/cty"
@@ -25,6 +34,7 @@ type UnparsedVariableValue interface {
 	// If error diagnostics are returned, the resulting value may be invalid
 	// or incomplete.
 	ParseVariableValue(evalCtx *hcl.EvalContext, mode modconfig.VariableParsingMode) (*terraform.InputValue, tfdiags.Diagnostics)
+	ParseVariableValueToType(evalCtx *hcl.EvalContext, mode modconfig.VariableParsingMode, targetType cty.Type) (*terraform.InputValue, tfdiags.Diagnostics)
 }
 
 // ParseVariableValues processes a map of unparsed variable values by
@@ -61,8 +71,27 @@ func ParseVariableValues(evalContext *hcl.EvalContext, inputValueUnparsed map[st
 		} else {
 			mode = modconfig.VariableParseLiteral
 		}
-
-		val, valDiags := unparsedVal.ParseVariableValue(evalContext, mode)
+		var val *terraform.InputValue
+		var valDiags tfdiags.Diagnostics
+		// if the variable is a connection, use special case logic to parse the value
+		// - handling connection strings and pipes workspace handle
+		if declared && config.IsConnectionType() {
+			val, valDiags = parseConnectionVariableValue(evalContext, unparsedVal, mode)
+		} else {
+			// Find type of variable and parse the value
+			var ctyType cty.Type
+			for _, pv := range publicVariables {
+				if pv.ShortName == name {
+					ctyType = pv.Type
+					break
+				}
+			}
+			if ctyType == cty.NilType {
+				val, valDiags = unparsedVal.ParseVariableValue(evalContext, mode)
+			} else {
+				val, valDiags = unparsedVal.ParseVariableValueToType(evalContext, mode, ctyType)
+			}
+		}
 		diags = diags.Append(valDiags)
 		if valDiags.HasErrors() {
 			continue
@@ -161,6 +190,71 @@ func ParseVariableValues(evalContext *hcl.EvalContext, inputValueUnparsed map[st
 	return ret, diags
 }
 
+// this function handles the case that the value for a connection variable is being set with a connection string or cloud workspace identifier
+func parseConnectionVariableValue(evalContext *hcl.EvalContext, unparsedVal UnparsedVariableValue, mode modconfig.VariableParsingMode) (*terraform.InputValue, tfdiags.Diagnostics) {
+	var ctyVal cty.Value
+	var err error
+	if unparsedString, ok := unparsedVal.(UnparsedVariableValueString); ok {
+		str := unparsedString.Raw()
+		// is the value a workspace handle?
+		if steampipeconfig.IsPipesWorkspaceIdentifier(str) {
+			// verify the cloud token was provided
+			cloudToken := viper.GetString(constants.ArgPipesToken)
+			if cloudToken == "" {
+				return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to get pipes metadata", error_helpers.MissingCloudTokenError().Error())}
+			}
+
+			pipesMetadata, err := pipes.GetPipesMetadata(context.Background(), str, cloudToken)
+			if err != nil {
+				return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to get pipes metadata", err.Error())}
+			}
+			// create new  connection
+			c := connection.NewSteampipePgConnection(str, hcl.Range{}).(*connection.SteampipePgConnection)
+			c.ConnectionString = &pipesMetadata.ConnectionString
+			ctyVal, err = c.CtyValue()
+			if err != nil {
+				return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to get connection value", err.Error())}
+			}
+			return &terraform.InputValue{
+				Value:      ctyVal,
+				SourceType: terraform.ValueFromCLIArg,
+			}, nil
+
+		} else if backend.HasBackend(str) {
+			// is the value a connection string?
+			var c connection.PipelingConnection
+			switch {
+			case backend.IsDuckDBConnectionString(str):
+				c = connection.NewDuckDbConnection("command_arg", hcl.Range{}).(*connection.DuckDbConnection)
+				c.(*connection.DuckDbConnection).ConnectionString = &str
+			case backend.IsPostgresConnectionString(str):
+				c = connection.NewPostgresConnection("command_arg", hcl.Range{})
+				c.(*connection.PostgresConnection).ConnectionString = &str
+			case backend.IsMySqlConnectionString(str):
+				c = connection.NewMysqlConnection("command_arg", hcl.Range{})
+				c.(*connection.MysqlConnection).ConnectionString = &str
+			case backend.IsSqliteConnectionString(str):
+				c = connection.NewSqliteConnection("command_arg", hcl.Range{})
+				c.(*connection.SqliteConnection).ConnectionString = &str
+
+			default:
+				return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to get connection value", fmt.Sprintf("Invalid connection string: %s", str))}
+			}
+			ctyVal, err = c.CtyValue()
+			if err != nil {
+				return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to get connection value", err.Error())}
+			}
+
+			return &terraform.InputValue{
+				Value:      ctyVal,
+				SourceType: terraform.ValueFromCLIArg,
+			}, nil
+		}
+	}
+	// fall back to normal parsing
+	return unparsedVal.ParseVariableValue(evalContext, mode)
+}
+
 func getUndeclaredVariableError(name string, variablesMap *modconfig.ModVariableMap) string {
 	// is this a qualified variable?
 	if len(strings.Split(name, ".")) == 1 {
@@ -186,6 +280,31 @@ func getUndeclaredVariableError(name string, variablesMap *modconfig.ModVariable
 
 type UnparsedInteractiveVariableValue struct {
 	Name, RawValue string
+}
+
+func (v UnparsedInteractiveVariableValue) ParseVariableValueToType(evalCtx *hcl.EvalContext, mode modconfig.VariableParsingMode, targetType cty.Type) (*terraform.InputValue, tfdiags.Diagnostics) {
+	var diags tfdiags.Diagnostics
+	val, valDiags := mode.Parse(evalCtx, v.Name, v.RawValue)
+	diags = diags.Append(valDiags)
+	if diags.HasErrors() {
+		return nil, diags
+	}
+
+	if val.Type() == cty.String {
+		valTarget, err := hclhelpers.CoerceStringToGoBasedOnCtyType(val.AsString(), targetType)
+		if err != nil {
+			return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to coerce value", err.Error())}
+		}
+
+		val, err = hclhelpers.ConvertInterfaceToCtyValue(valTarget)
+		if err != nil {
+			return nil, tfdiags.Diagnostics{tfdiags.Sourceless(tfdiags.Error, "Failed to convert value", err.Error())}
+		}
+	}
+	return &terraform.InputValue{
+		Value:      val,
+		SourceType: terraform.ValueFromInput,
+	}, diags
 }
 
 func (v UnparsedInteractiveVariableValue) ParseVariableValue(evalCtx *hcl.EvalContext, mode modconfig.VariableParsingMode) (*terraform.InputValue, tfdiags.Diagnostics) {