+++ modconfig/mod_tree_item_impl.go
@@ -6,6 +6,7 @@ import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/turbot/pipe-fittings/cty_helpers"
 	"github.com/turbot/pipe-fittings/printers"
+	"github.com/turbot/pipe-fittings/schema"
 	"github.com/zclconf/go-cty/cty"
 	"golang.org/x/exp/maps"
 )
@@ -89,9 +90,15 @@ func (b *ModTreeItemImpl) GetDatabase() *string {
 	if b.Database != nil {
 		return b.Database
 	}
+
+	// if we have a parent, ask for its database
+	// (stop when we get to the mod - the mod database property has lower precedence)
 	if len(b.parents) > 0 {
-		return b.GetParents()[0].GetDatabase()
+		if parent := b.GetParents()[0]; parent.BlockType() != schema.BlockTypeMod {
+			return parent.GetDatabase()
+		}
 	}
+
 	return nil
 }
 
@@ -100,8 +107,12 @@ func (b *ModTreeItemImpl) GetSearchPath() []string {
 	if len(b.SearchPath) != 0 {
 		return b.SearchPath
 	}
+	// if we have a parent, ask for its search path
+	// (stop when we get to the mod - the mod database property has lower precedence)
 	if len(b.parents) > 0 {
-		return b.GetParents()[0].GetSearchPath()
+		if parent := b.GetParents()[0]; parent.BlockType() != schema.BlockTypeMod {
+			return parent.GetSearchPath()
+		}
 	}
 
 	return nil
@@ -112,9 +123,14 @@ func (b *ModTreeItemImpl) GetSearchPathPrefix() []string {
 	if len(b.SearchPathPrefix) != 0 {
 		return b.SearchPathPrefix
 	}
-	if len(b.GetParents()) > 0 {
-		return b.GetParents()[0].GetSearchPathPrefix()
+	// if we have a parent, ask for its search path prefix
+	// (stop when we get to the mod - the mod database property has lower precedence)
+	if len(b.parents) > 0 {
+		if parent := b.GetParents()[0]; parent.BlockType() != schema.BlockTypeMod {
+			return parent.GetSearchPath()
+		}
 	}
+
 	return nil
 }