+++ modconfig/variable.go
@@ -2,7 +2,6 @@ package modconfig
 
 import (
 	"fmt"
-
 	"github.com/hashicorp/hcl/v2"
 	typehelpers "github.com/turbot/go-kit/types"
 	"github.com/turbot/pipe-fittings/cty_helpers"
@@ -14,8 +13,6 @@ import (
 	"github.com/zclconf/go-cty/cty"
 	"github.com/zclconf/go-cty/cty/convert"
 )
-
-// TODO check DescriptionSet - still required?
 
 // Variable is a struct representing a Variable resource
 type Variable struct {
@@ -181,4 +178,8 @@ func (v *Variable) CtyValue() (cty.Value, error) {
 // These variables are not added to the eval context, but instead are resolved at execution time
 func (v *Variable) IsLateBinding() bool {
 	return IsLateBindingType(v.Type)
+}
+
+func (p *Variable) IsConnectionType() bool {
+	return IsConnectionType(p.Type)
 }